Relocate RPCErr to ripple/protocol
Adds STCurrency.h to libxrpl for Clio
Fixes PR #4885 botched by force push
