+++ src/ripple/app/paths/PathRequest.cpp
@@ -28,8 +28,8 @@
 #include <ripple/basics/Log.h>
 #include <ripple/beast/core/LexicalCast.h>
 #include <ripple/core/Config.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/UintTypes.h>
 
 #include <ripple/rpc/impl/Tuning.h>
