+++ src/ripple/protocol/impl/BuildInfo.cpp
@@ -33,7 +33,7 @@ namespace BuildInfo {
 //  and follow the format described at http://semver.org/
 //------------------------------------------------------------------------------
 // clang-format off
-char const* const versionString = "2.2.0-b1"
+char const* const versionString = "2.2.0-b2"
 // clang-format on
 
 #if defined(DEBUG) || defined(SANITIZER)
