+++ src/ripple/rpc/handlers/AMMInfo.cpp
@@ -20,9 +20,9 @@
 #include <ripple/app/misc/AMMUtils.h>
 #include <ripple/json/json_value.h>
 #include <ripple/ledger/ReadView.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/AMMCore.h>
 #include <ripple/protocol/Issue.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/rpc/Context.h>
 #include <ripple/rpc/impl/RPCHelpers.h>
 #include <grpcpp/support/status.h>
