+++ src/ripple/rpc/handlers/CrawlShards.cpp
@@ -19,10 +19,10 @@
 
 #include <ripple/app/main/Application.h>
 #include <ripple/basics/RangeSet.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/nodestore/DatabaseShard.h>
 #include <ripple/overlay/Overlay.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/resource/Fees.h>
 #include <ripple/rpc/Context.h>
