+++ src/ripple/rpc/handlers/GetCounts.cpp
@@ -26,10 +26,10 @@
 #include <ripple/basics/UptimeClock.h>
 #include <ripple/json/json_value.h>
 #include <ripple/ledger/CachedSLEs.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/nodestore/Database.h>
 #include <ripple/nodestore/DatabaseShard.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/Context.h>
 #include <ripple/shamap/ShardFamily.h>
