+++ src/ripple/rpc/handlers/NoRippleCheck.cpp
@@ -21,8 +21,8 @@
 #include <ripple/app/misc/LoadFeeTrack.h>
 #include <ripple/app/paths/TrustLine.h>
 #include <ripple/ledger/ReadView.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/TxFlags.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/Context.h>
