+++ src/ripple/rpc/handlers/ServerInfo.cpp
@@ -22,8 +22,8 @@
 #include <ripple/app/reporting/P2pProxy.h>
 #include <ripple/json/json_value.h>
 #include <ripple/json/json_writer.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/LedgerFormats.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/SField.h>
 #include <ripple/protocol/TER.h>
 #include <ripple/protocol/TxFormats.h>
