+++ src/ripple/rpc/handlers/Tx.cpp
@@ -24,9 +24,9 @@
 #include <ripple/app/misc/Transaction.h>
 #include <ripple/app/rdb/RelationalDatabase.h>
 #include <ripple/basics/ToString.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
 #include <ripple/protocol/NFTSyntheticSerializer.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/CTID.h>
 #include <ripple/rpc/Context.h>
