+++ src/ripple/rpc/handlers/TxHistory.cpp
@@ -25,8 +25,8 @@
 #include <ripple/core/DatabaseCon.h>
 #include <ripple/core/Pg.h>
 #include <ripple/core/SociDB.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/resource/Fees.h>
 #include <ripple/rpc/Context.h>
