+++ src/ripple/rpc/handlers/Unsubscribe.cpp
@@ -19,8 +19,8 @@
 
 #include <ripple/app/misc/NetworkOPs.h>
 #include <ripple/basics/Log.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/Context.h>
 #include <ripple/rpc/Role.h>
