+++ src/ripple/rpc/handlers/ValidationCreate.cpp
@@ -18,8 +18,8 @@
 //==============================================================================
 
 #include <ripple/basics/Log.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/protocol/Seed.h>
 #include <ripple/protocol/jss.h>
 #include <ripple/rpc/Context.h>
