+++ src/ripple/rpc/handlers/Validators.cpp
@@ -19,8 +19,8 @@
 
 #include <ripple/app/main/Application.h>
 #include <ripple/app/misc/ValidatorList.h>
-#include <ripple/net/RPCErr.h>
 #include <ripple/protocol/ErrorCodes.h>
+#include <ripple/protocol/RPCErr.h>
 #include <ripple/rpc/Context.h>
 
 namespace ripple {
